Fix checkpoint handling, validation, and training robustness
- Atomic checkpoint saves to prevent corruption on kill (kermttrainer)
- load_checkpoint returns (model, state) to enable optimizer/scheduler
  resume for interrupted finetune jobs
- Remove attn_out and fine_tune_coff from checkpoint-loaded model args
  to prevent model size blowup from old checkpoints (attn_out default
  128→8)
- Add num_tasks to model args for blinded test data prediction
- Default ffn_hidden_size to hidden_size when not specified
- Support blinded test data: skip loss/metrics, save predictions only
- Initialize TensorBoard writer before training loop
- Reset random seeds per fold in cross_validate
- Prevent split_data from deleting vocab files in output directory
- Fix split_data file count ceiling division
- Auto-detect GPU topology and configure NCCL P2P for pretrain DDP
- Validate SMILES in MolGraph and skip invalid SMILES in vocab building
- Fix lr_coff float type and TorchVocab return type hint

+++ kermt/util/scheduler.py
@@ -58,7 +58,7 @@ class NoamLR(_LRScheduler):
         self.init_lr = np.array([init_lr] * self.num_lrs)
         self.max_lr = np.array([max_lr] * self.num_lrs)
         self.final_lr = np.array([final_lr] * self.num_lrs)
-        self.lr_coff = np.array([1] * self.num_lrs)
+        self.lr_coff = np.array([1.0] * self.num_lrs, dtype=float)
         self.fine_tune_param_idx = fine_tune_param_idx
         self.lr_coff[self.fine_tune_param_idx] = fine_tune_coff
